Moved Elevator commands into Elevator folder.
